+++ changelog.md
@@ -3,6 +3,10 @@
 Not all commits are going to be explained here.
 
 ### 200427 - 1 [0.0.1]
+- Added, now you can write pure JavaScript code between %% and %%
+- You can now change the variables value
+
+### 200427 - 1 [0.0.1]
 - Fixed some compression bug
 - Changed temp to final
 - Added var and flex type of variables